+++ src/index.ts
@@ -2,8 +2,8 @@ export { fastifyApolloDrainPlugin } from "./drain-plugin.js";
 export { fastifyApolloHandler } from "./handler.js";
 export { fastifyApollo } from "./plugin.js";
 export {
-  ApolloFastifyContextFunctionArgument,
   ApolloFastifyContextFunction,
+  ApolloFastifyContextFunctionArgument,
   ApolloFastifyHandlerOptions,
   ApolloFastifyPluginOptions,
 } from "./types.js";